Update reclass_raster function
Function now handles differences between raster values of the tif file and reclass input table by introducing NAs for values in the raster not found in the input table and not using values of the input table not found in the raster.
Note: that all changes in the bat and sh file are due to an old dev_reclass raster branch and all updates are based on main branch.